Se modulariza el sistema
