+++ server/app/templates/youtube.html
@@ -99,7 +99,7 @@
                     <div class="carousel-inner">
                     {%for i in range(1, 5)%}
                         <div class="carousel-item {% if i == 1 %}active{% endif %}">
-                            <img src="https://logia.buenaventuracadiz.cl/bcp/image/photos/bcp_{{i}}.jpeg" alt="Pronto acá un carrusel de fotos de Logia" class="d-block w-100">
+                            <img src="https://logia.buenaventuracadiz.cl/bcp/image/bcp_{{i}}.jpeg" alt="Pronto acá un carrusel de fotos de Logia" class="d-block w-100">
                         </div>
                     {%endfor%}
                     </div>
